refactor all saliency tests
